safari-600.7-branch/LayoutTests/media/track/in-band/track-in-band-srt-mkv-cues-added-once.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../../media-file.js></script>
        <script src=../../video-test.js></script>
        <script src=../../in-band-tracks.js></script>
    </head>
    <body onload="testCuesAddedOnce('../../content/counting-subtitled-srt.mkv', 'subtitles')">
        <video controls></video>
        <p>Check that we don't have duplicate cues after seeking backwards.</p>
    </body>
</html>
