LayoutTests/media/track/in-band/track-in-band-srt-mkv-cues-added-once-expected.txt

Check that we don't have duplicate cues after seeking backwards.

EVENT(canplaythrough)

** Setting track 1 to showing
RUN(inbandTrack1 = video.textTracks[0])
RUN(inbandTrack1.mode = 'showing')
RUN(video.play())
EXPECTED (inbandTrack1.cues.length > '0') OK
RUN(video.pause())
RUN(video.currentTime = 0)
RUN(video.play())

EVENT(seeked)
EXPECTED (inbandTrack1.cues.length > '0') OK
RUN(video.pause())
Test all cues are equal OK
END OF TEST
